+++ orc8r/cloud/go/services/metricsd/servicers/southbound/servicer_southbound_test.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package servicers_test
+package southbound
 
 import (
 	"context"
@@ -19,22 +19,26 @@ import (
 	"flag"
 	"strconv"
 	"testing"
-	"time"
 
 	"github.com/golang/glog"
 	dto "github.com/prometheus/client_model/go"
+	prometheusProto "github.com/prometheus/client_model/go"
 	"github.com/stretchr/testify/assert"
 
 	configurator_test_init "magma/orc8r/cloud/go/services/configurator/test_init"
 	"magma/orc8r/cloud/go/services/configurator/test_utils"
 	device_test_init "magma/orc8r/cloud/go/services/device/test_init"
 	"magma/orc8r/cloud/go/services/metricsd/exporters"
-	"magma/orc8r/cloud/go/services/metricsd/servicers"
+	"magma/orc8r/cloud/go/services/metricsd/test_common"
 	tests "magma/orc8r/cloud/go/services/metricsd/test_common"
 	"magma/orc8r/cloud/go/services/metricsd/test_init"
 	"magma/orc8r/cloud/go/services/orchestrator/obsidian/models"
 	"magma/orc8r/lib/go/metrics"
 	"magma/orc8r/lib/go/protos"
+)
+
+var (
+	testLabels = []*prometheusProto.LabelPair{{Name: tests.MakeStrPtr("labelName"), Value: tests.MakeStrPtr("labelValue")}}
 )
 
 type testMetricExporter struct {
@@ -43,12 +47,6 @@ type testMetricExporter struct {
 	// error to return
 	retErr error
 }
-
-const (
-	MetricName = protos.MetricName_process_virtual_memory_bytes
-	LabelName  = protos.MetricLabelName_result
-	LabelValue = "success"
-)
 
 // Set verbosity so we can capture exporter error logging
 var _ = flag.Set("vmodule", "*=2")
@@ -70,13 +68,38 @@ func (e *testMetricExporter) Submit(metrics []exporters.MetricAndContext) error 
 
 func (e *testMetricExporter) Start() {}
 
+// Set verbosity so we can capture exporter error logging
+var _ = flag.Set("vmodule", "*=2")
+
+func TestMetricsContainerToMetricAndContexts(t *testing.T) {
+	testFamily := tests.MakeTestMetricFamily(prometheusProto.MetricType_GAUGE, 1, testLabels)
+	container := protos.MetricsContainer{
+		GatewayId: "gw1",
+		Family:    []*prometheusProto.MetricFamily{testFamily},
+	}
+
+	metricAndContext := metricsContainerToMetricAndContexts(&container, "testNetwork", "gw1")
+
+	assert.Equal(t, 1, len(metricAndContext))
+	ctx := metricAndContext[0].Context
+	family := metricAndContext[0].Family
+	assert.NotNil(t, ctx.AdditionalContext)
+	assert.Equal(t, "gw1", ctx.AdditionalContext.(*exporters.GatewayMetricContext).GatewayID)
+	assert.Equal(t, "testNetwork", ctx.AdditionalContext.(*exporters.GatewayMetricContext).NetworkID)
+
+	labels := family.GetMetric()[0].Label
+	assert.Equal(t, 3, len(labels))
+	assert.True(t, tests.HasLabel(labels, metrics.NetworkLabelName, "testNetwork"))
+	assert.True(t, tests.HasLabel(labels, metrics.GatewayLabelName, "gw1"))
+	assert.True(t, tests.HasLabel(labels, testLabels[0].GetName(), testLabels[0].GetValue()))
+}
 func TestCollect(t *testing.T) {
 	device_test_init.StartTestService(t)
 	configurator_test_init.StartTestService(t)
 
 	e := &testMetricExporter{}
 	test_init.StartNewTestExporter(t, e)
-	srv := servicers.NewMetricsControllerServer()
+	srv := NewMetricsControllerServer()
 
 	// Create test network
 	networkID := "metricsd_servicer_test_network"
@@ -88,9 +111,9 @@ func TestCollect(t *testing.T) {
 	ctx := id.NewContextWithIdentity(context.Background())
 	test_utils.RegisterGateway(t, networkID, gatewayID, &models.GatewayDevice{HardwareID: gatewayID})
 
-	name := strconv.Itoa(int(MetricName))
-	key := strconv.Itoa(int(LabelName))
-	value := LabelValue
+	name := strconv.Itoa(int(test_common.MetricName))
+	key := strconv.Itoa(int(test_common.LabelName))
+	value := test_common.LabelValue
 	float := 1.0
 	int_val := uint64(1)
 	counter_type := dto.MetricType_COUNTER
@@ -204,7 +227,7 @@ func TestCollectMismatchedGateway(t *testing.T) {
 
 	e := &testMetricExporter{}
 	test_init.StartNewTestExporter(t, e)
-	srv := servicers.NewMetricsControllerServer()
+	srv := NewMetricsControllerServer()
 
 	// Create test network
 	networkID := "metricsd_servicer_test_network"
@@ -219,9 +242,9 @@ func TestCollectMismatchedGateway(t *testing.T) {
 	// Mismatched gateway ID
 	mismatchgatewayID := "2876171d-bf38-4254-b4da-71a713954029"
 
-	name := strconv.Itoa(int(MetricName))
-	key := strconv.Itoa(int(LabelName))
-	value := LabelValue
+	name := strconv.Itoa(int(test_common.MetricName))
+	key := strconv.Itoa(int(test_common.LabelName))
+	value := test_common.LabelValue
 	float := 1.0
 	counter_type := dto.MetricType_COUNTER
 	counters := protos.MetricsContainer{
@@ -250,63 +273,4 @@ func TestCollectMismatchedGateway(t *testing.T) {
 	assert.True(t, tests.HasLabel(e.queue[0].Labels(), metrics.NetworkLabelName, networkID))
 	assert.True(t, tests.HasLabel(e.queue[0].Labels(), metrics.GatewayLabelName, gatewayID))
 
-}
-
-func TestConsume(t *testing.T) {
-	metricsChan := make(chan *dto.MetricFamily)
-	e := &testMetricExporter{}
-
-	test_init.StartNewTestExporter(t, e)
-	srv := servicers.NewMetricsControllerServer()
-
-	go srv.ConsumeCloudMetrics(metricsChan, "Host_name_place_holder")
-	fam1 := "test1"
-	fam2 := "test2"
-	go func() {
-		metricsChan <- &dto.MetricFamily{Name: &fam1, Metric: []*dto.Metric{{}}}
-		metricsChan <- &dto.MetricFamily{Name: &fam2, Metric: []*dto.Metric{{}}}
-	}()
-	time.Sleep(time.Second)
-	assert.Equal(t, 2, len(e.queue))
-}
-
-func TestPush(t *testing.T) {
-	device_test_init.StartTestService(t)
-	configurator_test_init.StartTestService(t)
-
-	e := &testMetricExporter{}
-	test_init.StartNewTestExporter(t, e)
-	srv := servicers.NewCloudMetricsControllerServer()
-
-	// Create test network
-	networkID := "metricsd_servicer_test_network"
-	test_utils.RegisterNetwork(t, networkID, "Test Network Name")
-
-	metricName := "test_metric"
-	value := 8.2
-	testLabel := &protos.LabelPair{Name: "labelName", Value: "labelValue"}
-	timestamp := int64(123456)
-
-	protoMet := protos.PushedMetric{
-		MetricName:  metricName,
-		Value:       value,
-		TimestampMS: timestamp,
-		Labels:      []*protos.LabelPair{testLabel},
-	}
-	pushedMetrics := protos.PushedMetricsContainer{
-		NetworkId: networkID,
-		Metrics:   []*protos.PushedMetric{&protoMet},
-	}
-
-	_, err := srv.Push(context.Background(), &pushedMetrics)
-	assert.NoError(t, err)
-	assert.Equal(t, 1, len(e.queue))
-	assert.Equal(t, metricName, e.queue[0].Name())
-	assert.Equal(t, 2, len(e.queue[0].Labels()))
-	assert.Equal(t, testLabel.Name, *e.queue[0].Labels()[0].Name)
-	assert.Equal(t, testLabel.Value, *e.queue[0].Labels()[0].Value)
-	assert.Equal(t, metrics.NetworkLabelName, *e.queue[0].Labels()[1].Name)
-	assert.Equal(t, networkID, *e.queue[0].Labels()[1].Value)
-	assert.Equal(t, timestamp, e.queue[0].TimestampMs())
-	assert.Equal(t, strconv.FormatFloat(value, 'f', -1, 64), e.queue[0].Value())
 }
